feat[dace]: Modified `gt_simplify()` (#1647)
Before the pass was just calling the native DaCe version. Now it will
not call the `PromoteScalarToSymbol` and `ConstantPropagation` pass.
This is because the lowering sometimes has to change between a symbol
and a scalar and back and back again and so on. Furthermore, it looks
like these passes have problems with that, so we excluded them.

This is a temporary solution, at the end, it might be feasible or good
to run the full simplify pass.