feat[dace]: Custom SDFG inline pass (#1649)
Added a custom pass for inlining SDFG. The function builds upon the traditional inline pass (`InlineSDFG`), however, before it is run some cleaning steps are preformed (`PruneConectors` and `PruneSymbols`) which increase the likelihood that the inlining can be done. The cost is that state fusing is performed.

Also the `gt_simplify()` function is modified, instead of the build in behaviour it will only run the GT4Py specific one.
This behaviour can not be changed.