feat[next]: AOT toolchain (#1545)
## Description

Partial implementation of https://hackmd.io/3EMvJ7zJQPepDd18qDU3sg

Missing:
- Integrate transforms and compilation steps into one workflow. This PR
does prepare the ground for such a design change however.

### Background:
The `gt4py.next` machinery to transform DSL functions into executable
code has so far had unrestricted access to the arguments the user wished
to pass to the executable function, including full fields and their
contents, as well as connectivity tables. In other words, only JIT
compilation was possible, even though in principle AOT compilation
should also work.

In principle this should not be necessary, and for most of the reusable,
pluggable steps in this machinery this access was restricted in a
previous PR. However, the restrictions were introduced on a per-step
level to make the changes easier to integrate.

In order to get pre-compiled programs, `icon4py` has used a workaround
to store the result of the first JIT compilation.

### Changes
This PR makes introduces the restriction that the machinery should only
have access to the types of field and scalar arguments, and everything
but the connectivity tables for connectivityes on a toolchain level.
With the exception that one single ITIR pass still requires access to
connectivity tables, however it has been tested that nothing else does.

This also enabled the following design changes:
- The field operator / program stages are now decoupled from the args
and kwargs, which means less of them are required
- The transformation steps could be simplified, because they only need
to deal with the part of the data they actually require and an adapter
can ensure interoperability
- "optional" workflow steps have been replaced with step orders that
depend on input type (see `next.otf.workflow.MultiWorkflow` and
`next.backend.Transforms` for details.
- `next.backend.Backend` now requires a `ModularExecutor` and will call
it's `.executor.otf_workflow` directly and not `.executor.__call__`.
This is what prepares for unifying transform and compile workflows. This
also means that any custom code in `.executor.__call__` is not going to
be called by backends. For new `ProgramExecutors` it is highly
recommended to put all such code into their `.otf_workflow.decorate`
step. The alternative is to only be usable with a custom backend.
- `decorator.FrozenProgram` JIT compiles on first call and subsequently
skips the whole toolchain. It is accessible via
`@gtx.program(frozen=True, backend=...) or `decorator.Program.freeze()`.
It is in principle also compatible with field operators.

### Required steps before merging
- [x] test with `icon4py` (passed:
https://github.com/C2SM/icon4py/pull/513) and add required changes

### Future steps
It would be convenient to create a nice API to automatically create
compile time arguments from type annotations to DSL functions. But this
should probably be done in a follow-up PR.

---------

Co-authored-by: Edoardo Paone <edoardo.paone@cscs.ch>
Co-authored-by: Till Ehrengruber <till.ehrengruber@cscs.ch>