fix[next][dace]: enable subdomain in let-statements (#1668)
This PR fixes the lowering of let-statements with fields defined on
sub-domain. The previous implementation was allocating dace array with
the exact domain of the computation, eventually not starting from index
0. This required the use of offsets in dace arrays to compensate for the
start index. There were two problems with this: 1) in the lowering from
GTIR to SDFG because the array offsets were not correctly propagated
across nested SDFGs for let-statements; and 2) the use of array offsets
caused issue to SDFG-inlining pass. In order to reduce the overall
complexity, this PR proposes to remove the usage of array offsets and
therefore always allocate temporary fields starting from index 0.