refactor[next]: Cleanup toolchain (#1660)
Over time a significant amount of technical debt in the toolchain
workflows accumulated resulting in hardly readable code, slowed progress
in refactoring efforts and long review cycles. This PR streamlines some
of the interfaces and makes existing technical debt more visible (either
by TODOs or refactorings).

- Bugfix to make `gt4py.next.backend.Transforms` actually work from
`itir.Program | itir.FencilDefinition` (by skipping `past_to_itir` step
in that case)
- Rename `AOTProgram` type alias to `CompilableProgram` (was
inconsistent before)
- Fix backend type annotations. In many cases `ppi.ProgramExecutor` was
used even though a backend was required.
- Remove entire `ProgramExecutor` protocol. This protocol was not
properly used anywhere anymore.
- Remove left-over argument canonicalization in `Program` with embedded
execution and instead only validate the arguments.
- For some reason there was a lot of code duplication in `decorator.py`
in the form of
  ```python
  if self.backend is not None and self.backend.transforms is not None:
      return self.backend.transforms.<SOME CODE>
  return next_backend.DEFAULT_TRANSFORMS.<SOME CODE>
  ```
  which has been cleaned up into
  ```python
  return next_backend._frontend_transforms.<SOME CODE>
  ```
The `self.backend.transforms is not None` condition has been dropped.
According to the type annotation this could never occur, but `NoBackend`
in the tests required this behavior. Instead `NoBackend` now initializes
the backend properly using the default transforms and the technical debt
is explicitly documented with a TODO at the place it occurs.
- Remove `CompileTimeArg` class. This class looked like an actual value,
but wasn't. In many places it was used it was immediately converted to a
`TypeSpec`. This was quite confusing as it appeared like values were
passed around, while actually it was types only. By using `TypeSpec`
directly it is clear what information is available & actually used.
Similarly `SizeArg` was also removed and just replaced by a
`ScalarType`.
- Removal of `ModularExecutor`. This class served two purposes:
- The `name` attribute was used to distinguish backends. The `name` is
now part of the backends directly avoiding an unnecessary indirection.
- `__call__` This method duplicated `backend.Backend.__call__` and
additionally dropped all keyword arguments (e.g. `column_axis`) except
for the offset provider. The method was only used in the iterator tests
via `FendefDispatcher`. To make this explicit this was moved to
`FendefDispatcher` which uses the backend directly now.
Removal of the class allowed to remove additional code-duplication in
`roundtrip.py`.


Open issues:
- `gt4py.next.backend.Backend.executor`: The class and `executor` &
`transforms` attribute should be renamed to better reflect what they are
doing.
  - `Backend` -> `Toolchain`
  - `transforms` -> `frontend_transforms`
  - `executor` -> `backend_transforms`
While this is a relatively simple change it touches many lines and would
make review of the PR unnecessarily hard. Additionally it would break
Icon4Py. Will be tackled in a new PR.ü