[next]: Fix inline lambda pass opcount preserving option (#1687)
In #1531 the `itir.Node` class got a `type` attribute, that until now
contributed to the hash computation of all nodes. As such two
`itir.SymRef` with the same `id`, but one with a type inferred and one
without (i.e. `None`) got a different hash value. Consequently the
`inline_lambda` pass did not recognize them as a reference to the same
symbol and erroneously inlined the expression even with
`opcount_preserving=True`. This PR fixes the hash computation, such that
again `node1 == node2` implies `hash(node1) == hash(node2)`.