feat[next][dace]: GTIR-to-DaCe lowering of map-reduce with skip values (#1694)
This PR extends the solution for map-reduce provided in #1683 with the
support for connectivity tables with skip values.

The field definition is extended with a `local_offset` attribute that
stores the offset provider used to build the values in the local
dimension. In case the local dimension is built by the `neighbors`
expression, the `local_offset` corresponds to the offset provider used
to access the neighbor dimension. Since this information is carried
along the data itself, whenever the data is accessed it is also possible
to access the corresponding offset provider and check whether the
neighbor index is valid or if there is a skip value. For local
dimensions already present in the program argument, this information is
retrieved from the field domain (enabled in new test case).

The data is accessed in the `map_` and `reduce` expressions. Here it is
now possible to check for skip values. Therefore, the main objective of
this PR is the lowering of map-reduce with skip values.

A secondary objective is to pave the road to simplify the lowering
logic, by getting rid of the `reduce_identity` value. The current
approach is propagate the `reduce_identity` value while visiting the
arguments to `reduce` expressions. By introducing `local_offset`, the
argument visitor will return the information needed to implement
`reduce` of local values in presence of skip values.