refactor[cartesian]: Replace `is_start_state=` with `is_start_block=` when buiding the SDFG for DaCe (#1709)
When building an SDFG for DaCe it is sometimes necessary to specify the
starting point of the graph. In the past, `is_start_state=` was used for
this. In never versions of Dace, `is_start_state` has been deprecated
and replaced with `is_start_block`. This PR replaces the two occurrences
in the `cartesian/` folder and removes warnings generated during test
runs. It looks like the `next/` folder has already been cleaned.