fix[next][dace]: Fix for nested SDFG outer data descriptor (#1726)
Fixes a bug in lowering of let-statements. The lambda expression of a
let-statement is lowered to a nested SDFG. The result data produced in
the nested SDFG is written to temporary data allocated in the parent
SDFG. The previous lowering was directly using a copy of the inner data
descriptor for the outer data. The bug is that some symbols for array
shape and strides might not be available in the parent SDFG, so we have
to apply the symbol mapping on the outer data descriptor.

The test case `test_gtir_let_lambda_with_cond` was modified to trigger
this bug and verify the fix.