fix[next]: Fix type preservation in CSE (#1736)
The common subexpression elimination uses typing information to decide
what expressions can be extracted. However, while extracting it creates
new nodes and uses the inline lambda pass, which did not preserve the
types. This was observed in PMAP and is fixed in this PR on a best
effort basis. Creating a minimal reproducible example is hard and since
multiple of us are considering making typing information an integral
part of the IR, e.g. by attaching the computation to the node instead of
having a separate pass, which would solve the problem automatically no
tests have been written.