feat[next]: GTIR embedded and GTFN temporaries with new lowering (#1648)
Use new lowering for GTIR embedded, and GTFN. Only the dace iterator
backend continues to use the old lowering.

Changes:
- Use GTIR lowering for all backends except for dace
- Old lowering and transformations only used in dace backend
- workflows defined in
[`gt4py.next.backend.LEGACY_TRANSFORMS`](https://github.com/GridTools/gt4py/pull/1648/files#diff-cf4385d02cbeacc310d4326350903b4cb6f9a61c7cd36dda162a5077ab8b8e86).
Variable can be removed in a cleanup PR.
- old `apply_common_transforms` in
[pass_manager_legacy.py](https://github.com/GridTools/gt4py/pull/1648/files#diff-db17bff48ac16ee75ff974a1b9af98e3cf0c850971ce9898aa55b635bb046b72).
Just a straight copy of the old function. No need to review, this is
just to avoid deleting until gtir based dace backend is ready.
- Re-add `symbolic_sizes` param. Was in temporary extraction, is now
part of the domain inference. In preparation of icon-exclaim tests

---------

Co-authored-by: Hannes Vogt <hannes@havogt.de>