fix[next]: Guard diskcache creation by file lock (#1745)
The disk cache used to cache compilation in the gtfn backend has a race
condition manifesting itself in `sqlite3.OperationalError: database is
locked` errors if multiple python processes try to initialize the
`diskcache.Cache` object concurrently. This PR fixes this by guarding
the object creation by a file-based lock in the same directory as the
database.

While this issue occurred frequently and was observed to be fixed on
distributed file systems, the lock does not guarantee correct behavior
in particular for accesses to the cache (beyond opening) since the
underlying SQLite database is unreliable when stored on an NFS based
file system. It does however ensure correctness of concurrent cache
accesses on a local file system. See more information here:

https://grantjenks.com/docs/diskcache/tutorial.html#settings
https://www.sqlite.org/faq.html#q5
https://github.com/tox-dev/filelock/issues/73
NFS safe locking:
https://gitlab.com/warsaw/flufl.lock
[Barry Warsaw / FLUFL Lock ·
GitLab](https://gitlab.com/warsaw/flufl.lock)