Changed the constant substitution implementation.
This is a bit more stable, but it it does not seem to work yet, if there is an access node that should be replaced.
Which was the actuall goal of this modification.
However, it solves some problems with multiple states, at least it is now a bit more consistent in that case.