bug[next]: Fix propagated symbols order stability across runs (#1788)
If statements in FOAST have a special `propagated_symbols` annex
attributed that contains all symbols that are defined and hence
available outside of the if. The compuation of these symbols is based on
a set operation whose order is not stable across runs. This PR fixes
that by sorting them. Since behavior changes across runs are hard to
test and we don't have any facilities for that no test is provided.