feature[next]: Non-tree-size-increasing collapse tuple on ifs (#1762)
Removing the tuple expressions across `if_` calls on ITIR has been a
pain point in the past. While the `PROPAGATE_TO_IF_ON_TUPLES` option of
the `CollapseTuplePass` works very reliably, the resulting increase in
the tree size has been prohibitive. With the refactoring to GTIR this
problem became much less pronounced, as we could restrict the
propagation to field-level, i.e., outside of stencils, but the tree
still grew exponentially in the number of references to boolean
arguments used inside `if_` conditions. This PR adds an additional
option `PROPAGATE_TO_IF_ON_TUPLES_CPS` to the `CollapseTuplePass`, which
is similar to the existing `PROPAGATE_TO_IF_ON_TUPLES`, but propagates
in the opposite direction, i.e. into the tree. This allows removal of
tuple expressions across `if_` calls without increasing the size of the
tree. This is particularly important for `if` statements in the
frontend, where outwards propagation can have devastating effects on the
tree size, without any gained optimization potential. For example
```
complex_lambda(if cond1
  if cond2
    {...}
  else:
    {...}
else
  {...})
```
is problematic, since `PROPAGATE_TO_IF_ON_TUPLES` would propagate, and
hence duplicate, `complex_lambda` three times, while we only want to get
rid of the tuple expressions inside of the `if_`s. Note that this
transformation is not mutually exclusive to `PROPAGATE_TO_IF_ON_TUPLES`.