This commit fixes the stride reconstruction during stride propagation.
NestedSDFG essentially allows to perform some slices, there are technically three chases:
- The data container on the inside has a smaller rank than the one on the outside, thus some dimensions were removed.
- The data container on the inside has the same rank than the one on the outside.
- The data container on the inside has a larger rank than the one on the outside, thus some dimensions were added.

The last case is not handled, as it does not happens in GT4Py.
Before, the first and second case were handled together, but it was realized that the second case was not implemented properly and it was added explicitly.