This commit fixes a bug in `DistributedBufferRelocator` that was observed in ICON4Py's `TestUpdateThetaAndExner` test.
In essence there was an `assert` that assumed that checked if this temporary was a sink node.
but, the code that finds all write backs was never excluding such cases, i.e. the temporaries that were selected might not be sink nodes in the state where they are defined.
The `assert` was not part of the original implementation and is not a requirement of the transformation, instead it was introduced by [PR#1799](https://github.com/GridTools/gt4py/pull/1799), that fixed some issues in the analysis of read write dependencies.

There are two solutions for this, either removing the `assert` or prune these kinds of temporaries.
After some consideration, it was realized that handling such cases will not lead to invalid SDFG, as long as the other restrictions on the global data are respected. For that reason the `assert` was removed.
However, we should thinking of doing something more intelligent in that case.