Updated the transformations to conform (more) with DaCe main.
The main issue with the change to DaCe main is that now hierarchical control flow blocks are used.
Although we do not (yet) generate them simplify might create them.
It started in the `DistributedBufferRelocator` which used the `AccessSets` analysis pass.
Which now had become unusable, because it merges all accesses in a hierarchical block together.
So we switched to the `AccessStates` pass, but now we have to scan the states.

However, the main issue was the `is_accessed_downstream()`, because we can no longer simply explore the state machine, we now depend on the `ReachableStates` pass.
The downside is that this changed the meaning of the `states_to_ignore`.
For now we will life with that but we have to address that in the future.