fix[next][dace]: make if_ always execute branch exclusively (#1846)
This PR reverts the change previously made in #1824. Lowering `if_`
expressions to tasklet is semantically wrong, from a dataflow
perspective. It causes segmentation faults in several stencils, that
rely on exclusive branch execution.

The source problem was that full array shape was passed into the nested
SDFG scope, which prevented map fusion in most cases. This PR extends
the lowering with the detection of simple iterator dereferencing,
without shifts: for this type of data access, only the local element is
moved into the nested SDFG. However, when shift is applied on the
iterator input (which typically happens in iterator view), the full
array shape is still passed.

This approach increases the optimization opportunities by enabling more
map fusion. At the same time, it keeps the `if_` semantics of exclusive
branch execution.