fix[dace][next]: Update MapFusion (#1857)
Last year, the then state of MapFusion from [DaCe
PR#1629](https://github.com/spcl/dace/pull/1629) was added to GT4Py, as
a temporary fix until the PR in DaCe is merged and parallel map fusion
has become available there. However, during that time the transformation
in the PR has evolved and improved and some of the bug that were fixed
are now appearing in GT4Py, for example
[PR#1850](https://github.com/GridTools/gt4py/pull/1850) and
[PR#1856](https://github.com/GridTools/gt4py/pull/1856).

Thus this PR updated the MapFusion transformation that is currently
inside GT4Py and replaces it with newest development version from DaCe.
Because we need it, and it was designed from the start to be that way,
it also adds parallel map fusion to the transformation. As before, this
transformation, currently fully located in `map_fusion_dace.py`, is only
kept inside the repo until DaCe has caught up to it.

The PR also introduces some additional memory layer that encapsulates
the DaCe transformation. Something that we have to deal with in the
long run and we currently do because other parts of the toolchain require it.

---------

Co-authored-by: edopao <edoardo16@gmail.com>