Increas test coverage for read after write stencils
Read after write acccess patterns are hard to work with because one
needs to keep track of whether or not this variable has already been
writen in the current "scope".

In the current bridge where every vertical loop is a library node
with just one Tasklet, the "scope" is simple. In the new bridge where
codeflow is exposed to DaCe, we get many smaller Tasklets and it is thus
"harder" to keep the "scope" updated.