Fixed an issue in the auto optimizer.
The problem was that in the auto optimizer the "set stride correctly" function was called after the "go to GPU transformation".
The GPU transformation will call `CopyToMap` transformation, this is needed such that we can set the GPU block size and their order.
However, the desicion if we call `CopyToMap` also depends on teh strides, so we need to handle them.

But there is still a problem, but it is a prblem with the strides.