feat[next][dace]: Added `MultiStateGlobalSelfCopyElimination` (#1890)
The transformation `MultiStateGlobalSelfCopyElimination` is very
similar to the `GT4PyGlobalSelfCopyElimination` but they target slightly
different cases. The transformation `GT4PyGlobalSelfCopyElimination`,
which is already there and was renamed to `SingleStateGlobalSelfCopyElimination`,
handles the pattern `(G) -> (T) -> (G)`, i.e. the global data `G` is copied into
the transient `T` is then immediately copied back into `G`.
Because of ADR-18 we know that this has no effect, because `G` is used
as input and output and must therefore be point wise, so `G[i, j]` in
the output can only be `G[i, j]` at the beginning.

The new transformation `MultiStateGlobalSelfCopyElimination` handles a
different case, it looks for patterns `(G) -> (T)` and `(T) -> (G)`,
which is essentially the same, but this time the definition of `T` and
the write back of `T` into `G` does not need to be in the same state.

In the long run, the two transformation should be combined.

---------

Co-authored-by: edopao <edoardo16@gmail.com>