refactor[next]: Global ordering relation of dimensions (#1847)
This PR introduces a global ordering relation of dimensions, replacing
the previous mechanism in promote_dims. The ordering relation is:
sorting first by `Dimension.kind` (`HORIZONTAL` < `VERTICAL` < `LOCAL`)
and then lexicographically by `Dimension.value`.

Reason: 
An as_fieldop call as emitted by the frontend has no domain, and
inferring the type of the domain was not possible, since no global
ordering relation of dimensions existed, e.g. for an as_fieldop
operating on a `Vertex` and a `K` field it was unclear if the dimensions
of the output were `(Vertex, K)` or (`K, Vertex)`, which lead to many
negative consequences in other places, that will be tackled in [PR
1853](https://github.com/GridTools/gt4py/pull/1853) and following ones.