feat[dace][next]: Added `CopyChainRemover` (#1901)
This PR adds the `CopyChainRemover` transformation.

The introduction of `concat_where` introduced a new pattern, which is
essentially a chain of copies.
As an example imagine the case that a domain is split into 3 subdomain.
The result on the first subdomain is stored in `T1`, the one of the
second in `T2` and the one for the third domain in `T3`.
`T1` and `T2` are then copied into `T4`, finally `T4` together with `T3`
are then copied into `T5`.
This transformation will remove `T1`, `T2`, `T3` and `T4` thus the
results will be written into `T5` directly.

There are some limitation, if we have the pattern `(A1) -> (A2)`, then
we eliminate `A1` only if:
- `A1` is fully read; this is to avoid some nasty adjustments of map
bounds.
- There can only be one connection between `A1` and `A2`.

The transformation was added twice to the simplify pass, which allows us
to mitigate DaCe [issue#1959](https://github.com/spcl/dace/issues/1959).

---------

Co-authored-by: edopao <edoardo16@gmail.com>